+++ example/android/app/src/main/kotlin/ffm/example/example/MainActivity.kt
@@ -1,0 +1,6 @@
+package ffm.example.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}