+++ CHANGELOG.md
@@ -2,3 +2,13 @@
 ### Added
 * FPage, FPageLogic: split UI and logic, and provide connection to each other.
 * FPipe: easy state management with auto disposal.
+
+## 1.0.1
+### Added
+* FPageLogic: page transition animation
+* FPipe: error pipe embedded, subscribe when using withTextEditingController
+* ff: logger, time
+* example project
+
+### Changed
+* FPipe: onUpdate follow the T generic type